+++ assets/css/bootstrap/_extra_bespoke.scss
@@ -44,3 +44,7 @@
     box-shadow: 0 0 0 0.25rem rgba($form-feedback-valid-color, 0.25) !important;
   }
 }
+
+.form-control[readonly] {
+  font-family: $font-family-monospace;
+}
